% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylog.R
\name{mylog}
\alias{mylog}
\title{A slow logarithm function}
\usage{
mylog(x)
}
\arguments{
\item{x}{a numeric vector of positive values}
}
\description{
A slow function wrapper that computes the logarithm of a vector through a for loop
}
